rename test files
