+++ gwbackupy/tests/test_gmail.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 from gwbackupy.gmail import Gmail
 from gwbackupy.helpers import random_string, encode_base64url
-from gwbackupy.tests.mock_storage import MockStorage
+from gwbackupy.tests.test_mock_storage import MockStorage
 from gwbackupy.tests.test_mock_gmail_service_wrapper import MockGmailServiceWrapper
